@@ -1,6 +1,6 @@
 module github.com/wal-g/wal-g
 
-go 1.17
+go 1.18
 
 require (
 	cloud.google.com/go/storage v1.8.0
